+++ docs/ChangeLog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Mounted file systems scanner.
 - Unix timestamp to ISO 8601 formatted string conversion function in the **Timer** class.
 - Optional **format** parameter in the Wall clock to date/time formatter function of the **Timer** class.
 - **Storage::truncate** function for file truncation.